Better support for generic beats and additional metadata
This commit makes `BeatsCodec` handle Beats events more generically.

Instead of handling Filebeat, Packetbeat, Metricbeat, etc. individually,
the codec now tries to discover the name of the beat and prefix all fields
found in the Beats event with that name, e. g. "type" in a Filebeat event
will become "filebeat_type" in the resulting message, and "process_id" in
a Winlogbeat event will become "winlogbeat_process_id" in the resulting
message.

The "facility" field will always be "beats" and the "beats_type" field
will contain the name of the Beat, e. g. "filebeat" for Filebeat and
"beat" for any other generic beat which doesn't have the respective
information in "@metadata".

These changes are breaking backward compatibility with messages
generated by previous versions of this plugin, so extractors and
pipeline rules will have to be adapted to the new message schemas.

Closes #25